// API.ts
export type AggregatedStats = {
  __typename: "AggregatedStats",
  createdAt: string,
  id: string,
  losses: number,
  metadata?: string | null,
  owner?: string | null,
  statsKey: string,
  statsType: string,
  totalGames: number,
  updatedAt?: string | null,
  winRate: number,
  wins: number,
};

export type Deck = {
  __typename: "Deck",
  className: string,
  createdAt?: string | null,
  deckName: string,
  id: string,
  owner?: string | null,
  updatedAt: string,
  userId: string,
};

export type MatchRecord = {
  __typename: "MatchRecord",
  createdAt: string,
  id: string,
  isFirstPlayer: boolean,
  isWin: boolean,
  myDeckId: string,
  opponentClass: string,
  opponentDeckType: string,
  owner?: string | null,
  recordedAt: string,
  updatedAt: string,
  userId: string,
};

export type User = {
  __typename: "User",
  createdAt?: string | null,
  email: string,
  id: string,
  owner?: string | null,
  updatedAt: string,
};

export type ModelAggregatedStatsFilterInput = {
  and?: Array< ModelAggregatedStatsFilterInput | null > | null,
  createdAt?: ModelStringInput | null,
  id?: ModelIDInput | null,
  losses?: ModelIntInput | null,
  metadata?: ModelStringInput | null,
  not?: ModelAggregatedStatsFilterInput | null,
  or?: Array< ModelAggregatedStatsFilterInput | null > | null,
  owner?: ModelStringInput | null,
  statsKey?: ModelStringInput | null,
  statsType?: ModelStringInput | null,
  totalGames?: ModelIntInput | null,
  updatedAt?: ModelStringInput | null,
  winRate?: ModelFloatInput | null,
  wins?: ModelIntInput | null,
};

export type ModelStringInput = {
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  beginsWith?: string | null,
  between?: Array< string | null > | null,
  contains?: string | null,
  eq?: string | null,
  ge?: string | null,
  gt?: string | null,
  le?: string | null,
  lt?: string | null,
  ne?: string | null,
  notContains?: string | null,
  size?: ModelSizeInput | null,
};
// ...
export type ModelIDInput = {
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  beginsWith?: string | null,
  between?: Array< string | null > | null,
  contains?: string | null,
  eq?: string | null,
  ge?: string | null,
  gt?: string | null,
  le?: string | null,
  lt?: string | null,
  ne?: string | null,
  notContains?: string | null,
  size?: ModelSizeInput | null,
};

export type ModelIntInput = {
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  between?: Array< number | null > | null,
  eq?: number | null,
  ge?: number | null,
  gt?: number | null,
  le?: number | null,
  lt?: number | null,
  ne?: number | null,
};

export type ModelFloatInput = {
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  between?: Array< number | null > | null,
  eq?: number | null,
  ge?: number | null,
  gt?: number | null,
  le?: number | null,
  lt?: number | null,
  ne?: number | null,
};

export type ModelAggregatedStatsConnection = {
  __typename: "ModelAggregatedStatsConnection",
  items:  Array<AggregatedStats | null >,
  nextToken?: string | null,
};

export type ModelDeckFilterInput = {
  and?: Array< ModelDeckFilterInput | null > | null,
  className?: ModelStringInput | null,
  createdAt?: ModelStringInput | null,
  deckName?: ModelStringInput | null,
  id?: ModelIDInput | null,
  not?: ModelDeckFilterInput | null,
  or?: Array< ModelDeckFilterInput | null > | null,
  owner?: ModelStringInput | null,
  updatedAt?: ModelStringInput | null,
  userId?: ModelIDInput | null,
};

export type ModelDeckConnection = {
  __typename: "ModelDeckConnection",
  items:  Array<Deck | null >,
  nextToken?: string | null,
};

export type ModelMatchRecordFilterInput = {
  and?: Array< ModelMatchRecordFilterInput | null > | null,
  createdAt?: ModelStringInput | null,
  id?: ModelIDInput | null,
  isFirstPlayer?: ModelBooleanInput | null,
  isWin?: ModelBooleanInput | null,
  myDeckId?: ModelIDInput | null,
  not?: ModelMatchRecordFilterInput | null,
  opponentClass?: ModelStringInput | null,
  opponentDeckType?: ModelStringInput | null,
  or?: Array< ModelMatchRecordFilterInput | null > | null,
  owner?: ModelStringInput | null,
  recordedAt?: ModelStringInput | null,
  updatedAt?: ModelStringInput | null,
  userId?: ModelIDInput | null,
};

export type ModelBooleanInput = {
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  eq?: boolean | null,
  ne?: boolean | null,
};

export type ModelMatchRecordConnection = {
  __typename: "ModelMatchRecordConnection",
  items:  Array<MatchRecord | null >,
  nextToken?: string | null,
};

export type ModelUserFilterInput = {
  and?: Array< ModelUserFilterInput | null > | null,
  createdAt?: ModelStringInput | null,
  email?: ModelStringInput | null,
  id?: ModelIDInput | null,
  not?: ModelUserFilterInput | null,
  or?: Array< ModelUserFilterInput | null > | null,
  owner?: ModelStringInput | null,
  updatedAt?: ModelStringInput | null,
};

export type ModelUserConnection = {
  __typename: "ModelUserConnection",
  items:  Array<User | null >,
  nextToken?: string | null,
};
// ...
export type GetAggregatedStatsQueryVariables = {
  id: string,
};

export type GetAggregatedStatsQuery = {
  getAggregatedStats?:  {
    __typename: "AggregatedStats",
    createdAt: string,
    id: string,
    losses: number,
    metadata?: string | null,
    owner?: string | null,
    statsKey: string,
    statsType: string,
    totalGames: number,
    updatedAt?: string | null,
    winRate: number,
    wins: number,
  } | null,
};

export type GetDeckQueryVariables = {
  id: string,
};

export type GetDeckQuery = {
  getDeck?:  {
    __typename: "Deck",
    className: string,
    createdAt?: string | null,
    deckName: string,
    id: string,
    owner?: string | null,
    updatedAt: string,
    userId: string,
  } | null,
};

export type GetMatchRecordQueryVariables = {
  id: string,
};

export type GetMatchRecordQuery = {
  getMatchRecord?:  {
    __typename: "MatchRecord",
    createdAt: string,
    id: string,
    isFirstPlayer: boolean,
    isWin: boolean,
    myDeckId: string,
    opponentClass: string,
    opponentDeckType: string,
    owner?: string | null,
    recordedAt: string,
    updatedAt: string,
    userId: string,
  } | null,
};

export type GetUserQueryVariables = {
  id: string,
};

export type GetUserQuery = {
  getUser?:  {
    __typename: "User",
    createdAt?: string | null,
    email: string,
    id: string,
    owner?: string | null,
    updatedAt: string,
  } | null,
};

export type ListAggregatedStatsQueryVariables = {
  filter?: ModelAggregatedStatsFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type ListAggregatedStatsQuery = {
  listAggregatedStats?:  {
    __typename: "ModelAggregatedStatsConnection",
    items:  Array< {
      __typename: "AggregatedStats",
      createdAt: string,
      id: string,
      losses: number,
      metadata?: string | null,
      owner?: string | null,
      statsKey: string,
      statsType: string,
      totalGames: number,
      updatedAt?: string | null,
      winRate: number,
      wins: number,
    } | null >,
    nextToken?: string | null,
  } | null,
};

export type ListDecksQueryVariables = {
  filter?: ModelDeckFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type ListDecksQuery = {
  listDecks?:  {
    __typename: "ModelDeckConnection",
    items:  Array< {
      __typename: "Deck",
      className: string,
      createdAt?: string | null,
      deckName: string,
      id: string,
      owner?: string | null,
      updatedAt: string,
      userId: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};

export type ListMatchRecordsQueryVariables = {
  filter?: ModelMatchRecordFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type ListMatchRecordsQuery = {
  listMatchRecords?:  {
    __typename: "ModelMatchRecordConnection",
    items:  Array< {
      __typename: "MatchRecord",
      createdAt: string,
      id: string,
      isFirstPlayer: boolean,
      isWin: boolean,
      myDeckId: string,
      opponentClass: string,
      opponentDeckType: string,
      owner?: string | null,
      recordedAt: string,
      updatedAt: string,
      userId: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};

export type ListUsersQueryVariables = {
  filter?: ModelUserFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type ListUsersQuery = {
  listUsers?:  {
    __typename: "ModelUserConnection",
    items:  Array< {
      __typename: "User",
      createdAt?: string | null,
      email: string,
      id: string,
      owner?: string | null,
      updatedAt: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};